pinned memoery

+++ MyGameEngine/src/MGE/Physics/PhysicsObject.h
@@ -50,7 +50,7 @@ namespace MGE {
 
 		inline Vec2 GetVelocity() const { return m_Velocity; }
 		// inline float GetMass() const { return m_mass; }
-		inline float GetUID() const { return m_uID; }
+		inline int GetUID() const { return m_uID; }
 		inline float GetRadius() const { return m_radius; }
 
 		// inline float GetGravity() const { return m_Gravity; }
